+++ modules/validators/builtin/include/nlsValidators_builtin_exports.h
@@ -25,16 +25,13 @@
 //=============================================================================
 #pragma once
 //=============================================================================
-#include "nlsError_manager_exports.h"
-#include "Messages.hpp"
-#include <string>
-//=============================================================================
-namespace Nelson {
-//=============================================================================
-NLSERROR_MANAGER_IMPEXP void
-Error(const std::wstring& msg, const std::wstring& id = L"", bool asCaller = false);
-NLSERROR_MANAGER_IMPEXP void
-Error(const std::string& msg, const std::string& id = "", bool asCaller = false);
-//=============================================================================
-} // namespace Nelson
+#ifdef _MSC_VER
+#ifdef NLSVALIDATORS_BUILTIN_EXPORTS
+#define NLSVALIDATORS_BUILTIN_IMPEXP __declspec(dllexport)
+#else
+#define NLSVALIDATORS_BUILTIN_IMPEXP __declspec(dllimport)
+#endif
+#else
+#define NLSVALIDATORS_BUILTIN_IMPEXP __attribute__((visibility("default")))
+#endif
 //=============================================================================